CP-IJ: Factor out multiprocessing to subclass
The ImageJ daemon doesn't need to know anything about CellProfiler and
in some cases it seems beneficial to have the multiprocess code isolated
(e.g. to define a main method which may be needed for some
 multiprocessing contexts).
