+++ cpij/__init__.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+"""
+CellStar package providing CellStar algorithm for segmentation of yeast cells in brightfield imagery.
+Date: 2013-2016
+Website: http://cellstar-algorithm.org/
+"""
+__author__ = 'Mark Hiner'
+__all__ = ["bridge"]
