+++ runimagejscript.py
@@ -4,8 +4,7 @@ from cellprofiler_core.image import Image
 from cellprofiler_core.module import Module
 from cellprofiler_core.preferences import ABSOLUTE_FOLDER_NAME
 from cellprofiler_core.setting.choice import Choice
-from cellprofiler_core.setting.text import Filename, Text, Directory, Alphanumeric, Integer, Float
-from cellprofiler_core.setting.text.alphanumeric.name.image_name import ImageName
+from cellprofiler_core.setting.text import Filename, Text, Directory, Alphanumeric
 from cellprofiler_core.constants.module import (
     IO_FOLDER_CHOICE_HELP_TEXT,
 )
@@ -23,16 +22,11 @@ from wx import Gauge
 from wx import Window
 from collections.abc import Iterable
 from threading import Thread
-import multiprocessing as mp
-from multiprocessing import Process
-from pathlib import Path
 from sys import platform
 import time
-import atexit
-import imagej
-import jpype
 import random
 import skimage.io
+import cpij.bridge as cpij
 
 __doc__ = """\
 RunImageJScript
@@ -64,37 +58,8 @@ ImageJ Scripting: https://imagej.net/Scripting
  
 """
 
-"""
-Constants for communicating with pyimagej
-"""
-PYIMAGEJ_KEY_COMMAND = "KEY_COMMAND"  # Matching value indicates what command to execute
-PYIMAGEJ_KEY_INPUT = "KEY_INPUT"  # Matching value is command-specific input object
-PYIMAGEJ_KEY_OUTPUT = "KEY_OUTPUT"  # Matching value is command-specific output object
-PYIMAGEJ_KEY_ERROR = "KEY_OUTPUT"  # Matching value is command-specific output object
-PYIMAGEJ_CMD_SCRIPT_PARSE = "COMMAND_SCRIPT_PARAMS"  # Parse a script file's parameters
-PYIMAGEJ_SCRIPT_PARSE_INPUTS = "SCRIPT_PARSE_INPUTS"  # Script input dictionary key
-PYIMAGEJ_SCRIPT_PARSE_OUTPUTS = "SCRIPT_PARSE_OUTPUTS"  # Script output dictionary key
-PYIMAGEJ_CMD_SCRIPT_RUN = "COMMAND_SCRIPT_RUN"  # Run a script
-PYIMAGEJ_SCRIPT_RUN_FILE_KEY = "SCRIPT_RUN_FILE_KEY"  # The script filename key
-PYIMAGEJ_SCRIPT_RUN_INPUT_KEY = "SCRIPT_RUN_INPUT_KEY"  # The script input dictionary key
-PYIMAGEJ_SCRIPT_RUN_CONVERT_IMAGES = "SCRIPT_RUN_CONVERT_IMAGES" # Whether images should be converted or not
-PYIMAGEJ_CMD_EXIT = "COMMAND_EXIT"  # Shut down the pyimagej daemon
-PYIMAGEJ_STATUS_CMD_UNKNOWN = "STATUS_COMMAND_UNKNOWN"  # Returned when an unknown command is passed to pyimagej
-PYIMAGEJ_STATUS_STARTUP_COMPLETE = "STATUS_STARTUP_COMPLETE"  # Returned after initial startup before daemon loop
-PYIMAGEJ_STATUS_STARTUP_FAILED = "STATUS_STARTUP_FAILED"  # Returned when imagej.init fails
-INPUT_CLASS = "INPUT"
-OUTPUT_CLASS = "OUTPUT"
-INIT_LOCAL = "Local"
-INIT_ENDPOINT = "Endpoint"
-INIT_LATEST = "Latest"
-
-global stop_progress_thread, imagej_process, to_imagej, from_imagej, init_display_string
+global stop_progress_thread
 stop_progress_thread = False # Used to control the display of progress graphics
-imagej_process = None  # A subprocess running pyimagej
-to_imagej = None  # Queue to pass data to pyimagej
-from_imagej = None  # queue to receive data from pyimagej
-init_display_string = None # Indicator string for how imagej was initialized
-
 
 class PyimagejError(EnvironmentError):
     """
@@ -143,241 +108,6 @@ def add_param_info_settings(group, param_name, param_type, param_class):
     )
 
 
-def convert_java_to_python_type(ij, return_value):
-    """
-    Helper method to convert ImageJ/Java values to python values that can be passed between via queues (pickled)
-
-    Parameters
-    ----------
-    ij : imagej.init(), required
-        ImageJ entry point
-    return_value : supported Java type, required
-        A value to convert from Java to python
-
-    Returns
-    ---------
-    An instance of a python type that can safely cross queues with the given value, or None if no valid type exists.
-    """
-    if return_value is None:
-        return None
-    return_class = return_value.getClass()
-    type_string = str(return_class.toString()).split()[1]
-
-    image_classes = (jpype.JClass('ij.ImagePlus'), jpype.JClass('net.imagej.Dataset'), jpype.JClass('net.imagej.ImgPlus'))
-
-    if type_string == "java.lang.String" or type_string == "java.lang.Character":
-        return str(return_value)
-    elif type_string == "java.lang.Integer" or type_string == "java.lang.Long" or type_string == "java.lang.Short":
-        return int(return_value)
-    elif type_string == "java.lang.Float" or type_string == "java.lang.Double":
-        return float(return_value)
-    elif type_string == "java.lang.Boolean":
-        if return_value:
-            return True
-        else:
-            return False
-    elif type_string == "java.lang.Byte":
-        return bytes(return_value)
-    elif bool((img_class for img_class in image_classes if issubclass(return_class, img_class))):
-        return ij.py.from_java(return_value)
-
-    # Not a supported type
-    return None
-
-
-def convert_java_type_to_setting(param_name, param_type, param_class):
-    """
-    Helper method to convert ImageJ/Java class parameter types to CellProfiler settings
-
-    Parameters
-    ----------
-    param_name : str, required
-        The name of the parameter
-    param_type : str, required
-        The Java class name describing the parameter type
-    param_class: str, required
-        One of {input_class} or {output_class}, based on the parameter use
-
-    Returns
-    ---------
-    One or more Settings of a type appropriate for param_type, named with param_name. Or None if no valid conversion exists.
-    """
-    type_string = param_type.split()[1]
-    img_strings = ("ij.ImagePlus", "net.imagej.Dataset", "net.imagej.ImgPlus")
-    if INPUT_CLASS == param_class:
-        param_label = param_name
-        if type_string == "java.lang.String":
-            return Alphanumeric(param_label, "")
-        if type_string == "java.lang.Character":
-            return Character(param_label, "")
-        elif type_string == "java.lang.Integer":
-            return Integer(param_label, 0, minval=-2 ** 31, maxval=((2 ** 31) - 1))
-        elif type_string == "java.lang.Long":
-            return Integer(param_label, 0, minval=-2 ** 63, maxval=((2 ** 63) - 1))
-        elif type_string == "java.lang.Short":
-            return Integer(param_label, 0, minval=-32768, maxval=32767)
-        elif type_string == "java.lang.Byte":
-            return Integer(param_label, 0, minval=-128, maxval=127)
-        elif type_string == "java.lang.Boolean":
-            return Boolean(param_label, 0)
-        elif type_string == "java.lang.Float":
-            return Float(param_label, minval=-2 ** 31, maxval=((2 ** 31) - 1))
-        elif type_string == "java.lang.Double":
-            return Float(param_label, minval=-2 ** 63, maxval=((2 ** 63) - 1))
-        elif type_string == "java.io.File":
-            param_dir = Directory(f'{param_label} directory', allow_metadata=False)
-            def set_directory_fn_app(path):
-                dir_choice, custom_path = param_dir.get_parts_from_path(path)
-                param_dir.join_parts(dir_choice, custom_path)
-            param_file = Filename(
-                param_label,
-                param_label,
-                get_directory_fn=param_dir.get_absolute_path,
-                set_directory_fn=set_directory_fn_app,
-                browse_msg=f'Choose {param_label} file'
-                )
-            return (param_dir, param_file)
-        elif bool((img_string for img_string in img_strings if type_string == img_string)):
-            return ImageSubscriber(param_label)
-    elif OUTPUT_CLASS == param_class:
-        if bool((img_string for img_string in img_strings if type_string == img_string)):
-            return ImageName("[OUTPUT, " + type_string + "] " + param_name, param_name, doc=
-            """
-            You may use this setting to rename the indicated output variable, if desired.
-            """
-                             )
-
-    return None
-
-
-def preprocess_script_inputs(ij, input_map, convert_images):
-    """Helper method to convert pythonic inputs to something that can be handled by ImageJ
-
-    In particular this is necessary for image inputs which won't be auto-converted by Jpype
-
-    Parameters
-    ----------
-    ij : imagej.init(), required
-        ImageJ entry point (from imagej.init())
-    input_map:
-        map of input names to values
-    convert_images:
-        boolean indicating if image inputs and outputs should be auto-converted to appropriate numeric types
-    """
-    for key in input_map:
-        if isinstance(input_map[key], Image):
-            cp_image = input_map[key].get_image()
-            # CellProfiler images are typically stored as floats which can cause unexpected results in ImageJ.
-            # By default, we convert to 16-bit int type, unless we're sure it's 8 bit in which case we use that.
-            if convert_images:
-                if input_map[key].scale==255:
-                    cp_image = skimage.img_as_ubyte(cp_image)
-                else:
-                    cp_image = skimage.img_as_uint(cp_image)
-            input_map[key] = ij.py.to_dataset(cp_image)
-
-
-def start_imagej_process(input_queue, output_queue, init_string):
-    """Python script to run when starting a new ImageJ process.
-    
-    All commands are initiated by adding a dictionary with a {pyimagej_key_command} entry to the {input_queue}. This
-    indicating which supported command should be executed. Some commands may take additional input, which is specified
-    in the dictionary with {pyimagej_key_input}.
-
-    Outputs are returned by adding a dictionary to the {output_queue} with the {pyimagej_key_output} key, or
-    {pyimagej_key_error} if an error occurred during script execution.
-    
-    Supported commands
-    ----------
-    {pyimagej_cmd_script_parse} : parse the parameters from an imagej script. 
-        inputs: script filename
-        outputs: dictionary with mappings
-            {pyimagej_script_parse_inputs} -> dictionary of input field name/value pairs
-            {pyimagej_script_parse_outputs} -> dictionary of output field name/value pairs
-    {pyimagej_cmd_script_run} : takes a set of named inputs from CellProfiler and runs the given imagej script
-        inputs: dictionary with mappings
-            {pyimagej_script_run_file_key} -> script filename
-            {pyimagej_script_run_input_key} -> input parameter name/value dictionary
-        outputs: dictionary containing output field name/value pairs
-    {pyimagej_cmd_exit} : shut down the pyimagej daemon.
-        inputs: none
-        outputs: none
-    
-    Return values
-    ----------
-    {pyimagej_status_cmd_unknown} : unrecognized command, no further output is coming
-
-    Parameters
-    ----------
-    input_queue : multiprocessing.Queue, required
-        This Queue will be polled for input commands to run through ImageJ
-    output_queue : multiprocessing.Queue, required
-        This Queue will be filled with outputs to return to CellProfiler
-    init_string : str, optional
-        This can be a path to a local ImageJ installation, or an initialization string per imagej.init(),
-        e.g. sc.fiji:fiji:2.1.0
-    """
-
-    ij = False
-
-    try:
-        if init_string:
-            # Attempt to initialize with the given string
-            ij = imagej.init(init_string)
-        else:
-            ij = imagej.init()
-    except jpype.JException as ex:
-        # Initialization failed
-        output_queue.put(PYIMAGEJ_STATUS_STARTUP_FAILED)
-        jpype.shutdownJVM()
-        return
-
-    script_service = ij.script()
-
-    # Signify output is complete
-    output_queue.put(PYIMAGEJ_STATUS_STARTUP_COMPLETE)
-
-    # Main daemon loop, polling the input queue
-    while True:
-        command_dictionary = input_queue.get()
-        cmd = command_dictionary[PYIMAGEJ_KEY_COMMAND]
-        if cmd == PYIMAGEJ_CMD_SCRIPT_PARSE:
-            script_path = command_dictionary[PYIMAGEJ_KEY_INPUT]
-            script_file = Path(script_path)
-            script_info = script_service.getScript(script_file)
-            script_inputs = {}
-            script_outputs = {}
-            for script_in in script_info.inputs():
-                script_inputs[str(script_in.getName())] = str(script_in.getType().toString())
-            for script_out in script_info.outputs():
-                script_outputs[str(script_out.getName())] = str(script_out.getType().toString())
-            output_queue.put({PYIMAGEJ_SCRIPT_PARSE_INPUTS: script_inputs,
-                              PYIMAGEJ_SCRIPT_PARSE_OUTPUTS: script_outputs})
-        elif cmd == PYIMAGEJ_CMD_SCRIPT_RUN:
-            script_path = (command_dictionary[PYIMAGEJ_KEY_INPUT])[PYIMAGEJ_SCRIPT_RUN_FILE_KEY]
-            script_file = Path(script_path)
-            input_map = (command_dictionary[PYIMAGEJ_KEY_INPUT])[PYIMAGEJ_SCRIPT_RUN_INPUT_KEY]
-            convert_types = (command_dictionary[PYIMAGEJ_KEY_INPUT])[PYIMAGEJ_SCRIPT_RUN_CONVERT_IMAGES]
-            preprocess_script_inputs(ij, input_map, convert_types)
-            script_out_map = script_service.run(script_file, True, input_map).get().getOutputs()
-            output_dict = {}
-            for entry in script_out_map.entrySet():
-                key = str(entry.getKey())
-                value = convert_java_to_python_type(ij, entry.getValue())
-                if value is not None:
-                    output_dict[key] = value
-
-            output_queue.put({PYIMAGEJ_KEY_OUTPUT: output_dict})
-        elif cmd == PYIMAGEJ_CMD_EXIT:
-            break
-        else:
-            output_queue.put({PYIMAGEJ_KEY_ERROR: PYIMAGEJ_STATUS_CMD_UNKNOWN})
-
-    # Shut down the daemon
-    ij.getContext().dispose()
-    jpype.shutdownJVM()
-
-
 class RunImageJScript(Module):
     """
     Module to run ImageJ scripts via pyimagej
@@ -407,9 +137,9 @@ class RunImageJScript(Module):
         self.set_notes([" ".join(module_explanation)])
 
         self.init_choice = Choice(
-            "Initialization type", [INIT_LOCAL, INIT_ENDPOINT, INIT_LATEST],
-            tooltips={INIT_LOCAL: "Use a local ImageJ/Fiji installation", INIT_ENDPOINT: "Specify a particular endpoint",
-                      INIT_LATEST: "Use the latest Fiji, downloading if needed."},
+            "Initialization type", [cpij.INIT_LOCAL, cpij.INIT_ENDPOINT, cpij.INIT_LATEST],
+            tooltips={cpij.INIT_LOCAL: "Use a local ImageJ/Fiji installation", cpij.INIT_ENDPOINT: "Specify a particular endpoint",
+                      cpij.INIT_LATEST: "Use the latest Fiji, downloading if needed."},
             doc="""\
 Note that initialization will only occur once per CellProfiler session! After initialization, these options will be
 locked for the remainder of the session.
@@ -421,9 +151,9 @@ Select the mechanism for initializing ImageJ:
 
 Note that any option besides {init_local} may result in a download of the requested components.
             """.format(
-                init_local=INIT_LOCAL,
-                init_endpoint=INIT_ENDPOINT,
-                init_latest=INIT_LATEST,
+                init_local=cpij.INIT_LOCAL,
+                init_endpoint=cpij.INIT_ENDPOINT,
+                init_latest=cpij.INIT_LATEST,
             ),
         )
 
@@ -445,10 +175,9 @@ If you choose to disable this function, your ImageJ script will need to account 
             """,
         )
 
-        global init_display_string
-        if init_display_string:
+        if cpij.init_display_string:
             # ImageJ thread is already running
-            self.initialized_method.set_value(init_display_string)
+            self.initialized_method.set_value(cpij.init_display_string)
 
         self.app_directory = Directory(
             "ImageJ directory", allow_metadata=False, doc="""\
@@ -515,51 +244,17 @@ Note: this must be done each time you change the script, before running the Cell
         or a version string.
         """
         choice = self.init_choice.get_value()
-        if choice == INIT_LATEST:
+        if choice == cpij.INIT_LATEST:
             return None
 
-        if choice == INIT_LOCAL:
+        if choice == cpij.INIT_LOCAL:
             init_string = self.app_directory.get_absolute_path()
             if platform == 'darwin':
                 init_string = path.join(init_string, self.app_file.value)
-        elif choice == INIT_ENDPOINT:
+        elif choice == cpij.INIT_ENDPOINT:
             init_string = self.endpoint_string.get_value()
 
         return init_string
-
-    def close_pyimagej(self):
-        """
-        Close the pyimagej daemon thread
-        """
-        global imagej_process, to_imagej
-        if imagej_process is not None:
-            to_imagej.put({PYIMAGEJ_KEY_COMMAND: PYIMAGEJ_CMD_EXIT})
-
-    def init_pyimagej(self):
-        """
-        Start the pyimagej daemon thread if it isn't already running.
-        """
-        self.initialization_failed = False
-        init_string = self.get_init_string()
-
-        global imagej_process, to_imagej, from_imagej, init_display_string
-        if imagej_process is None:
-            to_imagej = mp.Queue()
-            from_imagej = mp.Queue()
-            # TODO if needed we could set daemon=True
-            imagej_process = Process(target=start_imagej_process, name="PyImageJ Daemon",
-                                          args=(to_imagej, from_imagej, init_string,))
-            imagej_process.start()
-            result = from_imagej.get()
-            if result == PYIMAGEJ_STATUS_STARTUP_FAILED:
-                imagej_process = None
-                self.initialization_failed = True
-            else:
-                atexit.register(self.close_pyimagej)  # TODO is there a more CP-ish way to do this?
-                init_display_string = self.init_choice.get_value()
-                if init_display_string != INIT_LATEST:
-                    init_display_string += ": " + init_string
-                self.initialized_method.set_value(init_display_string)
 
     def clear_script_parameters(self):
         """
@@ -598,13 +293,24 @@ Note: this must be done each time you change the script, before running the Cell
 
         if not self.initialization_failed:
             self.parsed_params = True
+    
+    def init_pyimagej(self):
+        self.initialization_failed = False
+        init_string = self.get_init_string()
+        if cpij.init_pyimagej(init_string):
+            init_display_string = self.init_choice.get_value()
+            if init_display_string != cpij.INIT_LATEST:
+                init_display_string += ": " + init_string
+            self.initialized_method.set_value(init_display_string)
+        else:
+            self.initialization_failed = True
 
     def get_parameters_from_script(self):
         """
         Use PyImageJ to read header text from an ImageJ script and extract inputs/outputs, which are then converted to
         CellProfiler settings for this module
         """
-        global stop_progress_thread, imagej_process, to_imagej, from_imagej
+        global stop_progress_thread
         script_filepath = path.join(self.script_directory.get_absolute_path(), self.script_file.value)
 
         if not self.script_file.value or not path.exists(script_filepath):
@@ -614,25 +320,25 @@ Note: this must be done each time you change the script, before running the Cell
 
         # Start pyimagej if needed
         self.init_pyimagej()
-        if not imagej_process:
+        if not cpij.imagej_process:
             stop_progress_thread = True
             return
 
         # Tell pyimagej to parse the script parameters
-        to_imagej.put({PYIMAGEJ_KEY_COMMAND: PYIMAGEJ_CMD_SCRIPT_PARSE, PYIMAGEJ_KEY_INPUT: script_filepath})
+        cpij.to_imagej.put({cpij.PYIMAGEJ_KEY_COMMAND: cpij.PYIMAGEJ_CMD_SCRIPT_PARSE, cpij.PYIMAGEJ_KEY_INPUT: script_filepath})
 
-        ij_return = from_imagej.get()
+        ij_return = cpij.from_imagej.get()
 
         # Process pyimagej's output, converting script parameters to settings
-        if ij_return != PYIMAGEJ_STATUS_CMD_UNKNOWN:
-            input_params = ij_return[PYIMAGEJ_SCRIPT_PARSE_INPUTS]
-            output_params = ij_return[PYIMAGEJ_SCRIPT_PARSE_OUTPUTS]
+        if ij_return != cpij.PYIMAGEJ_STATUS_CMD_UNKNOWN:
+            input_params = ij_return[cpij.PYIMAGEJ_SCRIPT_PARSE_INPUTS]
+            output_params = ij_return[cpij.PYIMAGEJ_SCRIPT_PARSE_OUTPUTS]
 
-            for param_dict, settings_dict, io_class in ((input_params, self.script_input_settings, INPUT_CLASS),
-                                                        (output_params, self.script_output_settings, OUTPUT_CLASS)):
+            for param_dict, settings_dict, io_class in ((input_params, self.script_input_settings, cpij.INPUT_CLASS),
+                                                        (output_params, self.script_output_settings, cpij.OUTPUT_CLASS)):
                 for param_name in param_dict:
                     param_type = param_dict[param_name]
-                    next_setting = convert_java_type_to_setting(param_name, param_type, io_class)
+                    next_setting = cpij.convert_java_type_to_setting(param_name, param_type, io_class)
                     if next_setting is not None:
                         settings_dict[param_name] = next_setting
                         group = SettingsGroup()
@@ -669,19 +375,18 @@ Note: this must be done each time you change the script, before running the Cell
 
     def visible_settings(self):
         visible_settings = []
-        global imagej_process
 
         # Update the visible settings based on the selected initialization method
         # If ImageJ is already initialized we just want to report how it was initialized
         # Otherwise we show: a string entry for "endpoint", a directory chooser for "local" (and file chooser if on mac),
         # and nothing if "latest"
-        if not imagej_process:
+        if not cpij.imagej_process:
             visible_settings += [self.init_choice]
             input_type = self.init_choice.get_value()
             # ImageJ is not initialized yet
-            if input_type == INIT_ENDPOINT:
+            if input_type == cpij.INIT_ENDPOINT:
                 visible_settings += [self.endpoint_string]
-            elif input_type == INIT_LOCAL:
+            elif input_type == cpij.INIT_LOCAL:
                 visible_settings += [self.app_directory]
                 if platform == 'darwin':
                     visible_settings += [self.app_file]
@@ -722,7 +427,7 @@ Note: this must be done each time you change the script, before running the Cell
             param_name = setting_values[i - 2]
             param_type = setting_values[i - 1]
             io_class = setting_values[i]
-            setting = convert_java_type_to_setting(param_name, param_type, io_class)
+            setting = cpij.convert_java_type_to_setting(param_name, param_type, io_class)
             # account for remover, name, type and io_class
             i -= 4
             # account for the number of values in this setting
@@ -734,9 +439,9 @@ Note: this must be done each time you change the script, before running the Cell
             group.append("remover", RemoveSettingButton("", "Remove this variable", self.script_parameter_list, group))
             add_param_info_settings(group, param_name, param_type, io_class)
             loaded_settings.append(group)
-            if INPUT_CLASS == io_class:
+            if cpij.INPUT_CLASS == io_class:
                 self.script_input_settings[param_name] = setting
-            elif OUTPUT_CLASS == io_class:
+            elif cpij.OUTPUT_CLASS == io_class:
                 self.script_output_settings[param_name] = setting
             settings_count -= 1
 
@@ -767,7 +472,7 @@ Note: this must be done each time you change the script, before running the Cell
                 self.script_file
             )
 
-        if self.init_choice.get_value() == INIT_LOCAL:
+        if self.init_choice.get_value() == cpij.INIT_LOCAL:
             app_path = self.get_init_string()
             if not path.exists(app_path):
                 raise ValidationError(
@@ -776,14 +481,13 @@ Note: this must be done each time you change the script, before running the Cell
                 )
 
     def validate_module_warnings(self, pipeline):
-        global imagej_process
         """Warn user if the specified FIJI executable directory is not found, and warn that a copy of FIJI will be downloaded"""
         warn_msg = "Please note: any initialization method except \"Local\", a new Fiji may be downloaded"
         " to your machine if cached dependencies not found."
         init_type = self.init_choice.get_value()
-        if init_type != INIT_LOCAL:
+        if init_type != cpij.INIT_LOCAL:
             # The component we attach the error to depends on if initialization has happened or not
-            if not imagej_process:
+            if not cpij.imagej_process:
                 raise ValidationError(warn_msg, self.init_choice)
             else:
                 raise ValidationError(warn_msg + " If re-initialization is required, please restart CellProfiler.",
@@ -820,16 +524,16 @@ Note: this must be done each time you change the script, before running the Cell
                 script_inputs[name] = setting.get_value()
 
         # Start the script
-        to_imagej.put({PYIMAGEJ_KEY_COMMAND: PYIMAGEJ_CMD_SCRIPT_RUN, PYIMAGEJ_KEY_INPUT:
-            {PYIMAGEJ_SCRIPT_RUN_FILE_KEY: script_filepath,
-             PYIMAGEJ_SCRIPT_RUN_INPUT_KEY: script_inputs,
-             PYIMAGEJ_SCRIPT_RUN_CONVERT_IMAGES: self.convert_types.value}
+        cpij.to_imagej.put({cpij.PYIMAGEJ_KEY_COMMAND: cpij.PYIMAGEJ_CMD_SCRIPT_RUN, cpij.PYIMAGEJ_KEY_INPUT:
+            {cpij.PYIMAGEJ_SCRIPT_RUN_FILE_KEY: script_filepath,
+             cpij.PYIMAGEJ_SCRIPT_RUN_INPUT_KEY: script_inputs,
+             cpij.PYIMAGEJ_SCRIPT_RUN_CONVERT_IMAGES: self.convert_types.value}
                             })
 
         # Retrieve script output
-        ij_return = from_imagej.get()
-        if ij_return != PYIMAGEJ_STATUS_CMD_UNKNOWN:
-            script_outputs = ij_return[PYIMAGEJ_KEY_OUTPUT]
+        ij_return = cpij.from_imagej.get()
+        if ij_return != cpij.PYIMAGEJ_STATUS_CMD_UNKNOWN:
+            script_outputs = ij_return[cpij.PYIMAGEJ_KEY_OUTPUT]
             for name in self.script_output_settings:
                 output_key = self.script_output_settings[name].get_value()
                 output_value = script_outputs[name]
